Completed MVP, add new service img-gen-node, bug fixes
