+++ discord-bot-node/.env.example
@@ -1,6 +1,5 @@
-API_URL=dsb_php
-API_PORT=80
+API_URL=
+API_PORT=
 
 DISCORD_BOT_TOKEN=
 DISCORD_BOT_CLIENT_ID=
-DISCORD_DEV_GUILD_IDS=
