+++ discord-bot-node/package.json
@@ -23,10 +23,12 @@
   "dependencies": {
     "discord.js": "^14.14.1",
     "dotenv": "^16.3.1",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "multer": "^1.4.5-lts.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.21",
+    "@types/multer": "^1.4.11",
     "ts-node": "^10.9.2",
     "tsc-watch": "^6.0.4",
     "tsx": "^4.7.1",
