+++ discord-bot-node/src/BFFServer/index.ts
@@ -1,15 +1,24 @@
 import express from 'express'
 import { Client } from 'discord.js'
+import multer from 'multer'
+import { Bot } from '../structs/Bot'
 
-const app = express()
-const port = 8989
+export default async (discordClient: Client, botInstance: Bot) => {
+    const app = express()
+    const upload = multer();
+    const port = 8989
 
-export default async (discordClient: Client) => {
-    app.post('/send-image', async (req:express.Request, res:express.Response) => {
-        const payload = await req.body.json()
-        console.log(payload);
-        // discordClient.sendMessageLogic
-        res.send(200)
+    app.post('/send-image', upload.single('image'), async (req: express.Request, res: express.Response) => {
+        const file = req.file
+
+        if (file) {
+            const [serverId, channelId, matchId] = file.originalname.split('_')
+            const buffer = file.buffer
+
+            botInstance.sendImage(channelId, buffer, file.originalname)
+        }
+
+        res.sendStatus(200)
     })
 
     app.listen(port, () => {
