+++ dota-api-laravel/.env.example
@@ -63,3 +63,5 @@ GRAPHQL_ENDPOINT=""
 GRAPHQL_CREDENTIALS=""
 GRAPHQL_AUTHENTICATION_HEADER="Authorization"
 GRAPHQL_AUTHENTICATION="bearer"
+
+IMG_GEN_NODE_ENDPOINT=""
