+++ dota-api-laravel/Dockerfile
@@ -16,5 +16,3 @@ COPY --chown=www-data:www-data . /var/www/html
 WORKDIR /var/www/html
 
 RUN composer install
-
-RUN php artisan migrate
