+++ dota-api-laravel/app/Console/Commands/UpdateSteamAccountsData.php
@@ -30,23 +30,26 @@ class UpdateSteamAccountsData extends Command
     public function handle()
     {
         $this->info('Command "steamAccounts:update-data" executed successfully.');
+        Log::info('Command "steamAccounts:update-data" executed successfully.');
 
         $chunkedIds = SteamAccount::select('id')->get()->pluck('id')->chunk(5);
 
         $chunkedIds->each(function (Collection $chunk) {
             $steamAccountsData = SteamAccountService::getSteamAccountsProfileData($chunk->toArray());
-
+            
             foreach ($steamAccountsData as $data) {
+                Log::info('New data from API   ', ['data' => $data]);
+
                 $lastMatchData = $data['matches'][0];
 
                 if (!$lastMatchData['statsDateTime']) {
-                    return;
+                    continue;
                 }
 
                 $steamAccount = SteamAccount::find($data['steamAccountId']);
 
                 if ($lastMatchData['id'] == $steamAccount->last_match_id) {
-                    return;
+                    continue;
                 }
 
                 $steamAccount->last_match_id = $data['matches'][0]['id'];
