+++ dota-api-laravel/config/api.php
@@ -15,6 +15,14 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Dota stratz api URL
+    |--------------------------------------------------------------------------
+    */
+
+    'dsb_img_gen_node' => env('IMG_GEN_NODE_ENDPOINT', ''),
+
+    /*
+    |--------------------------------------------------------------------------
     | Dota stratz api credentials
     |--------------------------------------------------------------------------
     */
