+++ dota-api-laravel/resources/views/style/main-style.blade.php
@@ -1,0 +1,18 @@
+<style>
+    body {
+        font-family: 'Roboto', sans-serif;
+        font-size: 20px;
+        width: 915px;
+        margin: 0;
+        background-color: transparent;
+        color: #f2f2f2;
+    }
+    
+
+    .content__wrapper {
+        display: flex;
+        flex-direction: column;
+        row-gap: 7px;
+    }
+
+</style>
