+++ dota-api-laravel/routes/web.php
@@ -18,3 +18,4 @@ use Illuminate\Support\Facades\DB;
 
 Route::get('/get-match-image', [DiscordServerController::class, 'getMatchImage']);
 Route::get('/get-player-image', [DiscordServerController::class, 'getPlayerImage']);
+Route::get('/get-finally-image', [DiscordServerController::class, 'getFoo']);
