+++ img-gen-node/Dockerfile.dev
@@ -1,0 +1,32 @@
+FROM node:slim
+
+# We don't need the standalone Chromium
+ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD true
+
+# Install Google Chrome Stable and fonts
+# Note: this installs the necessary libs to make the browser work with Puppeteer.
+RUN apt-get update && apt-get install gnupg wget -y && \
+    wget --quiet --output-document=- https://dl-ssl.google.com/linux/linux_signing_key.pub | gpg --dearmor > /etc/apt/trusted.gpg.d/google-archive.gpg && \
+    sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list' && \
+    apt-get update && \
+    apt-get install google-chrome-stable -y --no-install-recommends && \
+    rm -rf /var/lib/apt/lists/*
+
+# FROM public.ecr.aws/lambda/nodejs:14.2022.09.09.11
+# Create working directory
+WORKDIR /usr/src/app
+
+# Copy package.json
+COPY package.json ./
+
+# Install NPM dependencies for function
+RUN npm install
+
+# Copy handler function and tsconfig
+COPY server.js ./
+
+# Expose app
+EXPOSE 8090
+
+# Run app
+CMD ["npm", "run", "start:dev"]
