--- img-gen-python/requirements.txt
@@ -1,3 +1,0 @@
-fastapi==0.109.2
-html2image==2.0.4.3
-uvicorn==0.27.1
